add embeds-pack collators module + re-point test
